GP2-1208 - create a hook to make endpoint, debounce, save and error state resuable
