+++ package.json
@@ -63,8 +63,8 @@
     "eslint-plugin-import": "latest",
     "eslint-plugin-jest": "latest",
     "eslint-plugin-jsx-a11y": "latest",
-    "eslint-plugin-prettier": "latest",
-    "eslint-plugin-react": "latest",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-wc": "latest",
     "fetch-mock": "8.3.2",
     "file-loader": "^5.1.0",
@@ -81,7 +81,7 @@
     "postcss-preset-env": "6.7.0",
     "postcss-scss": "^3.0.4",
     "postcss-url": "8.0.0",
-    "prettier": "latest",
+    "prettier": "^2.2.1",
     "react-test-renderer": "16.12.0",
     "redux-devtools": "3.5.0",
     "resolve-url-loader": "^3.1.2",
